+++ javascript/Inits/LoadContent.js
@@ -1,7 +1,7 @@
 const morphdom = require('morphdom').default;
 
 window.libs['load::content'] = function (name) {
-    axios.post(window.load_content, {name: name})
+    axios.post(window.load_content + location.search, {name: name, _load_content: true})
         .then((response) => {
             if (response.data) {
                 morphdom(this.target, response.data.content ? `<div>${response.data.content}</div>` : "<div></div>");