+++ src/Middlewares/ApiMiddleware.php
@@ -84,8 +84,10 @@ class ApiMiddleware
 
         /** @var \Illuminate\Http\Response $response */
         $response = $next($request);
+
         //dd($response->exception);
         if (static::isApi()) {
+
 
             Auth::guard('admin')->logout();