+++ translations/uk/admin.php
@@ -317,4 +317,8 @@ return [
     'add_widget' => 'Додати віджет',
     'invalid_code' => 'Недійсний код!',
     'invalid_credentials' => 'Недійсні облікові дані!',
+    'private_notes' => 'Приватні нотатки',
+    'private_note_saved_successfully' => 'Приватна нотатка успішно збережена!',
+    'dashboard_saved' => 'Панель приладів збережена!',
+    'dashboard_not_found' => 'Панель приладів не знайдена!',
 ];